fixes yarn lint errors
