--- eslint.config.js
@@ -1,50 +1,0 @@
-// @ts-check
-const globals = require('globals');
-
-/** @type {import('eslint').Linter.FlatConfig[]} */
-module.exports = [
-  // Ignore patterns
-  {
-    ignores: [
-      'node_modules/**',
-      'dist/**',
-      '.nuxt/**',
-      'coverage/**',
-      '*.min.js',
-    ],
-  },
-  // Default config for all files
-  {
-    languageOptions: {
-      ecmaVersion: 2020,
-      sourceType: 'module',
-      globals: {
-        ...globals.browser,
-        ...globals.node,
-        ...globals.jest,
-      },
-      parserOptions: {
-        ecmaFeatures: {
-          jsx: true,
-        },
-      },
-    },
-    files: ['**/*.js', '**/*.ts'],
-    rules: {
-      'no-undef': 'error',
-      'no-unused-vars': 'warn',
-      'no-console': 'warn',
-    },
-  },
-  // Vue files
-  {
-    files: ['**/*.vue'],
-    languageOptions: {
-      parser: require('vue-eslint-parser'),
-      parserOptions: {
-        ecmaVersion: 2020,
-        sourceType: 'module',
-      },
-    },
-  },
-];
