+++ package.json
@@ -58,6 +58,7 @@
     "@types/geojson": "^7946.0.16",
     "@types/jest": "^29.5.14",
     "@types/node": "^22.13.15",
+    "@typescript-eslint/parser": "^8.29.0",
     "@vue/compiler-dom": "^3.5.13",
     "@vue/server-renderer": "^3.5.13",
     "@vue/server-test-utils": "^1.3.6",
@@ -72,6 +73,7 @@
     "eslint": "^9.0.0",
     "eslint-config-prettier": "^10.1.1",
     "eslint-plugin-html": "^8.1.2",
+    "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-nuxt": "^4.0.0",
     "eslint-plugin-vue": "^10.0.0",
     "globals": "^16.0.0",
